+++ .gitignore
@@ -10,5 +10,4 @@ npm-debug.log
 .DS_Store
 dist/
 .idea/
-.vscode/
 deeplearn-*.tgz
