+++ models/knn_image_classifier/knn_image_classifier.ts
@@ -15,8 +15,9 @@
  * =============================================================================
  */
 // tslint:disable-next-line:max-line-length
-import {Array1D, Array2D, Array3D, Model, NDArrayMath, Scalar, util} from 'deeplearn';
+import {Array1D, Array2D, Array3D, Model, NDArrayMath, Scalar} from 'deeplearn';
 import {SqueezeNet} from 'deeplearn-squeezenet';
+import * as model_util from '../util';
 
 export class KNNImageClassifier implements Model {
   private squeezeNet: SqueezeNet;
@@ -178,7 +179,7 @@ export class KNNImageClassifier implements Model {
     const knn = this.predict(image).asType('float32');
     const numExamples = this.getNumExamples();
     const kVal = Math.min(this.k, numExamples);
-    const topK = util.topK(await knn.data(), kVal);
+    const topK = model_util.topK(await knn.data(), kVal);
     knn.dispose();
     const topKIndices = topK.indices;
