+++ models/knn_image_classifier/package.json
@@ -5,8 +5,10 @@
   "main": "dist/knn_image_classifier/index.js",
   "unpkg": "dist/bundle.js",
   "types": "dist/knn_image_classifier/index.d.ts",
+  "peerDependencies": {
+    "deeplearn": "~0.5.0"
+  },
   "dependencies": {
-    "deeplearn": "latest",
     "deeplearn-squeezenet": "latest"
   },
   "repository": {
@@ -16,6 +18,7 @@
   "devDependencies": {
     "browserify": "~14.5.0",
     "browserify-shim": "~3.8.14",
+    "deeplearn": "~0.5.0",
     "tsify": "~3.0.3",
     "tslint": "~5.8.0",
     "typescript": "2.7.2",
