+++ models/squeezenet/squeezenet.ts
@@ -15,7 +15,8 @@
  * =============================================================================
  */
 // tslint:disable-next-line:max-line-length
-import {Array1D, Array3D, Array4D, CheckpointLoader, Model, NDArray, NDArrayMath, util} from 'deeplearn';
+import {Array1D, Array3D, Array4D, CheckpointLoader, Model, NDArray, NDArrayMath} from 'deeplearn';
+import * as model_util from '../util';
 import {IMAGENET_CLASSES} from './imagenet_classes';
 
 const GOOGLE_CLOUD_STORAGE_DIR =
@@ -174,7 +175,7 @@ export class SqueezeNet implements Model {
     const predictions = this.math.scope(() => {
       return this.math.softmax(logits).asType('float32');
     });
-    const topk = util.topK(await predictions.data(), topK);
+    const topk = model_util.topK(await predictions.data(), topK);
     predictions.dispose();
     const topkIndices = topk.indices;
     const topkValues = topk.values;
